requirements.txt

streamlit
requests
baidu-aip
speechrecognition
pygame
